collapsing all endpoint_* files into single files
i.e. collapsing beer_*.go all into beer.go... same for
event, brewery, guild, and search.
