+++ search.go
@@ -87,3 +87,90 @@ func (ss *SearchService) search(query string, q *SearchRequest, t searchType, da
 	}
 	return ss.c.Do(req, data)
 }
+
+// GeoPointUnit differentiates between miles and kilometers.
+type GeoPointUnit string
+
+// Units of measurement.
+const (
+	Miles      GeoPointUnit = "mi"
+	Kilometers GeoPointUnit = "km"
+)
+
+// GeoPointRequest contains options for specifying a geographic coordinate.
+type GeoPointRequest struct {
+	Latitude           float64      `url:"lat"` // Required
+	Longitude          float64      `url:"lng"` // Required
+	Radius             float64      `url:"radius,omitempty"`
+	Unit               GeoPointUnit `url:"unit,omitempty"`               // Default: mi
+	WithSocialAccounts string       `url:"withSocialAccounts,omitempty"` // Y/N
+	WithGuilds         string       `url:"withGuilds,omitempty"`         // Y/N
+	WithAlternateNames string       `url:"withAlternateNames,omitempty"` // Y/N
+}
+
+// GeoPoint searches for Locations near the geographic coordinate specified in the GeoPointRequest.
+// TODO: pagination??
+func (ss *SearchService) GeoPoint(q *GeoPointRequest) ([]Location, error) {
+	req, err := ss.c.NewRequest("GET", "/search/geo/point", q)
+	if err != nil {
+		return nil, err
+	}
+
+	geoPointResult := struct {
+		NumberOfPages int
+		CurrentPage   int
+		TotalResults  int
+		Data          []Location
+	}{}
+
+	err = ss.c.Do(req, &geoPointResult)
+	return geoPointResult.Data, err
+}
+
+// Style retrieves one or more Styles matching the given query string.
+// TODO: pagination??
+func (ss *SearchService) Style(query string, withDescriptions bool) ([]Style, error) {
+	q := struct {
+		Query            string `url:"q"`
+		WithDescriptions string `url:"withDescriptions,omitempty"`
+	}{Query: query}
+	if withDescriptions {
+		q.WithDescriptions = "Y"
+	}
+
+	req, err := ss.c.NewRequest("GET", "/search/style", &q)
+	if err != nil {
+		return nil, err
+	}
+
+	styleResponse := struct {
+		NumberOfPages int
+		CurrentPage   int
+		TotalResults  int
+		Data          []Style
+	}{}
+	err = ss.c.Do(req, &styleResponse)
+	return styleResponse.Data, err
+}
+
+// UPC retrieves one or more Beers matching the given Universal Product Code.
+// TODO: pagination??
+func (ss *SearchService) UPC(code uint64) ([]Beer, error) {
+	q := struct {
+		Code uint64 `url:"code"`
+	}{code}
+
+	req, err := ss.c.NewRequest("GET", "/search/upc", &q)
+	if err != nil {
+		return nil, err
+	}
+
+	upcResponse := struct {
+		NumberOfPages int
+		CurrentPage   int
+		TotalResults  int
+		Data          []Beer
+	}{}
+	err = ss.c.Do(req, &upcResponse)
+	return upcResponse.Data, err
+}
